Add InvokeArgs and InvokeResponse.
